requirements.txt
Flask>=2.3.0
Pillow>=10.1.0
Werkzeug>=2.3.0
